renaming tpro strategy
